Fix dtypes

+++ library/train_util.py
@@ -6312,7 +6312,7 @@ def get_noise_noisy_latents_and_timesteps(args, noise_scheduler, latents, fixed_
         t = timesteps.float() / max_timestep  # t ranges from 0 to 1
 
         if scaling_type == 'none':
-            scaling_factor = torch.ones_like(timesteps, dtype=torch.float32, device=latents.device)
+            scaling_factor = torch.ones_like(timesteps, dtype=latents.dtype, device=latents.device)
         elif scaling_type == 'linear':
             scaling_factor = t
         elif scaling_type == 'sine':
